Mise en place de policy et de request et début de l'interface

+++ app/Http/Controllers/ItemController.php
@@ -6,11 +6,16 @@ use Illuminate\Http\Request;
 use App\Models\Item;
 use Illuminate\Http\RedirectResponse;
 use Inertia\Inertia;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 
 class ItemController extends Controller
 {
+    use AuthorizesRequests;
+
     public function index(Request $request): \Inertia\Response
     {
+        $this->authorize('viewAny', Item::class);
+
         // Récupère la valeur de 'paginationMaxDisplay' depuis la requête, avec une valeur par défaut de 25
         $paginationMaxDisplay = (int) $request->input('paginationMaxDisplay', 25);
 
@@ -32,6 +37,8 @@ class ItemController extends Controller
 
     public function show(Item $item, Request $request): Item | RedirectResponse | Inertia
     {
+        $this->authorize('view', $item);
+
         $getView = (bool) $request->input('getView', false);
 
         if ($getView) {
@@ -44,11 +51,15 @@ class ItemController extends Controller
 
     public function create(): \Inertia\Response
     {
+        $this->authorize('create', Item::class);
+
         return Inertia::render('item.create');
     }
 
     public function store(Request $request): RedirectResponse
     {
+        $this->authorize('create', Item::class);
+
         $item = Item::create($request->all());
 
         return redirect()->route('item.show', ['item' => $item]);
@@ -56,6 +67,8 @@ class ItemController extends Controller
 
     public function edit(Item $item): \Inertia\Response
     {
+        $this->authorize('update', $item);
+
         return Inertia::render('item.edit', [
             'item' => $item
         ]);
@@ -63,6 +76,8 @@ class ItemController extends Controller
 
     public function update(Item $item, Request $request): RedirectResponse
     {
+        $this->authorize('update', $item);
+
         $item->update($request->all());
 
         return redirect()->route('item.show', ['item' => $item]);
@@ -70,7 +85,27 @@ class ItemController extends Controller
 
     public function delete(Item $item): RedirectResponse
     {
+        $this->authorize('delete', $item);
+
         $item->delete();
+
+        return redirect()->route('item.index');
+    }
+
+    public function forceDelete(Item $item): RedirectResponse
+    {
+        $this->authorize('forceDelete', $item);
+
+        $item->forceDelete();
+
+        return redirect()->route('item.index');
+    }
+
+    public function restore(Item $item): RedirectResponse
+    {
+        $this->authorize('restore', $item);
+
+        $item->restore();
 
         return redirect()->route('item.index');
     }
